scripts/system_improvements.py
"""
System Analysis & Improvement Recommendations
Analyzes current system and suggests enhancements
"""

import os
from pathlib import Path

def analyze_system():
    print("=" * 80)
    print("SMART RESUME ANALYZER - SYSTEM ANALYSIS")
    print("=" * 80)
    
    improvements = {
        "Critical": [],
        "High Priority": [],
        "Medium Priority": [],
        "Nice to Have": []
    }
    
    # Check current status
    print("\n[1] CURRENT SYSTEM STATUS")
    print("-" * 80)
    
    # Check models
    models_exist = all([
        os.path.exists("models/tfidf_vectorizer.pkl"),
        os.path.exists("models/doc2vec_model.model"),
        os.path.exists("models/random_forest_model.pkl")
    ])
    
    if models_exist:
        print("✅ All AI models trained and saved")
    else:
        print("❌ Some models missing")
        improvements["Critical"].append("Train missing AI models")
    
    # Check dashboard
    if os.path.exists("dashboard/streamlit_app.py"):
        print("✅ Dashboard exists")
    else:
        improvements["Critical"].append("Dashboard file missing")
    
    # Check data
    if os.path.exists("data/raw/UpdatedResumeDataSet.csv"):
        print("✅ Resume dataset available")
    else:
        improvements["Critical"].append("Resume dataset missing")
    
    print("\n[2] IDENTIFIED IMPROVEMENTS")
    print("-" * 80)
    
    # === CRITICAL IMPROVEMENTS ===
    improvements["Critical"].extend([
        # None if system is working
    ])
    
    # === HIGH PRIORITY IMPROVEMENTS ===
    improvements["High Priority"].extend([
        "Add experience extraction from resume text",
        "Add education level matching",
        "Improve model accuracy with more training data",
        "Add support for DOCX file format",
        "Add resume parsing for structured sections (Experience, Education, Skills)",
        "Implement caching to speed up repeat analyses"
    ])
    
    # === MEDIUM PRIORITY IMPROVEMENTS ===
    improvements["Medium Priority"].extend([
        "Add keyword highlighting in resume/job text",
        "Add PDF generation for analysis reports",
        "Add email notification for batch analysis",
        "Add user authentication for multi-user deployment",
        "Add database to store analysis history",
        "Add analytics dashboard (most requested skills, trends)",
        "Support for multiple languages",
        "Add ATS (Applicant Tracking System) compatibility score",
        "Add resume formatting suggestions",
        "Add salary prediction based on skills"
    ])
    
    # === NICE TO HAVE ===
    improvements["Nice to Have"].extend([
        "Add dark mode for dashboard",
        "Add export to JSON/XML formats",
        "Add API endpoints for integration",
        "Add Chrome extension for LinkedIn profiles",
        "Add mobile-responsive design",
        "Add skill trend analysis over time",
        "Add job market insights",
        "Add resume template suggestions",
        "Add interview question generator based on skills",
        "Add cover letter analyzer"
    ])
    
    # Print improvements
    for priority, items in improvements.items():
        if items:
            print(f"\n{priority.upper()}:")
            for i, item in enumerate(items, 1):
                print(f"  {i}. {item}")
    
    print("\n[3] RECOMMENDED QUICK WINS")
    print("-" * 80)
    
    quick_wins = [
        {
            "name": "Add DOCX Support",
            "effort": "Low",
            "impact": "High",
            "code": """
# In data_preprocessing.py, add:
from docx import Document

def extract_text_from_docx(self, docx_file):
    doc = Document(docx_file)
    return ' '.join([para.text for para in doc.paragraphs])
"""
        },
        {
            "name": "Add Experience Extraction",
            "effort": "Medium",
            "impact": "High",
            "code": """
# In data_preprocessing.py, add:
def extract_years_of_experience(self, text):
    import re
    patterns = [
        r'(\\d+)\\+?\\s*years?\\s*(?:of)?\\s*experience',
        r'experience\\s*:?\\s*(\\d+)\\+?\\s*years?'
    ]
    for pattern in patterns:
        match = re.search(pattern, text.lower())
        if match:
            return int(match.group(1))
    return 0
"""
        },
        {
            "name": "Add Result Caching",
            "effort": "Low",
            "impact": "Medium",
            "code": """
# In dashboard/streamlit_app.py, add:
import hashlib

@st.cache_data
def analyze_resume_cached(resume_hash, job_hash):
    # Your analysis code here
    return results
"""
        },
        {
            "name": "Add Keyword Highlighting",
            "effort": "Low",
            "impact": "Medium",
            "code": """
# In utils.py, add:
def highlight_keywords(text, keywords):
    for keyword in keywords:
        text = text.replace(
            keyword, 
            f'<mark style="background-color: yellow;">{keyword}</mark>'
        )
    return text
"""
        },
        {
            "name": "Add ATS Score",
            "effort": "Medium",
            "impact": "High",
            "code": """
# In feature_engineering.py, add:
def calculate_ats_score(self, resume_text):
    score = 100
    # Deduct points for:
    # - Tables (hard to parse)
    # - Images (can't read)
    # - Complex formatting
    # - Non-standard section names
    return score
"""
        }
    ]
    
    for i, qw in enumerate(quick_wins, 1):
        print(f"\n{i}. {qw['name']}")
        print(f"   Effort: {qw['effort']} | Impact: {qw['impact']}")
        print(f"   Implementation:")
        print(qw['code'])
    
    print("\n[4] PERFORMANCE IMPROVEMENTS")
    print("-" * 80)
    
    perf_improvements = [
        "Use multiprocessing for batch analysis (10x faster)",
        "Implement progressive loading for large files",
        "Add text truncation for very long documents",
        "Cache preprocessed text to avoid reprocessing",
        "Use incremental model training for updates",
        "Optimize Doc2Vec inference with GPU support"
    ]
    
    for i, imp in enumerate(perf_improvements, 1):
        print(f"  {i}. {imp}")
    
    print("\n[5] USER EXPERIENCE IMPROVEMENTS")
    print("-" * 80)
    
    ux_improvements = [
        "Add progress bar for long operations",
        "Add tooltips explaining each score",
        "Add example resumes/jobs for demo",
        "Add 'Save Analysis' button",
        "Add comparison view (before/after resume update)",
        "Add skill learning resources links",
        "Add resume templates download",
        "Add guided tour for first-time users"
    ]
    
    for i, imp in enumerate(ux_improvements, 1):
        print(f"  {i}. {imp}")
    
    print("\n[6] BUSINESS VALUE ADDITIONS")
    print("-" * 80)
    
    business_improvements = [
        "Add ROI calculator (time saved, cost per hire reduction)",
        "Add compliance checking (GDPR, equal opportunity)",
        "Add skill gap analysis for entire organization",
        "Add recruitment funnel analytics",
        "Add candidate pipeline management",
        "Add integration with HR systems (Workday, SAP)",
        "Add white-label customization for clients"
    ]
    
    for i, imp in enumerate(business_improvements, 1):
        print(f"  {i}. {imp}")
    
    print("\n[7] IMMEDIATE ACTION ITEMS")
    print("-" * 80)
    print("""
Based on current state, here are the TOP 5 improvements to implement:

1. ⭐ Add DOCX Support (1 hour)
   - Many resumes are in DOCX format
   - Easy to implement with python-docx
   - High user impact

2. ⭐ Add Experience Extraction (2 hours)
   - Extract years of experience from text
   - Use in ML model for better predictions
   - Improve accuracy significantly

3. ⭐ Add Result Caching (30 minutes)
   - Speed up repeat analyses
   - Better user experience
   - Simple Streamlit decorator

4. ⭐ Add ATS Compatibility Score (3 hours)
   - Help users make resumes ATS-friendly
   - Unique selling point
   - High business value

5. ⭐ Add Resume Section Parser (4 hours)
   - Extract structured sections (Skills, Experience, Education)
   - More accurate skill extraction
   - Enable section-by-section comparison

TOTAL TIME: ~10 hours for significant improvements
""")
    
    print("\n[8] TECHNICAL DEBT")
    print("-" * 80)
    
    tech_debt = [
        "Add unit tests for all modules",
        "Add integration tests for dashboard",
        "Add error logging to file",
        "Add monitoring/alerting for production",
        "Add API documentation",
        "Refactor large functions into smaller ones",
        "Add type hints throughout codebase",
        "Add configuration file (config.yaml)"
    ]
    
    for i, item in enumerate(tech_debt, 1):
        print(f"  {i}. {item}")
    
    print("\n" + "=" * 80)
    print("ANALYSIS COMPLETE")
    print("=" * 80)
    print("\nRecommendation: Start with the 5 immediate action items above.")
    print("These will provide the most value with minimal effort.")
    print("=" * 80)

if __name__ == "__main__":
    analyze_system()
